+++ packages/ritext/src/index.ts
@@ -9,3 +9,5 @@ export { Bold } from "./lib/extensions/Bold";
 export { Italic } from "./lib/extensions/Italic";
 export { Underline } from "./lib/extensions/Underline";
 export { Strike } from "./lib/extensions/Strike";
+export { Subscript } from "./lib/extensions/Subscript";
+export { Superscript } from "./lib/extensions/Superscript";